Created new module core.util.math and moved FT calculation functions from core.util.units Moved type checking and in_range functions from core.util.units to core.util.helpers. General code cleanup.
