+++ gui/magnet/magnet_gui.py
@@ -28,9 +28,7 @@ import pyqtgraph.exporters
 from qtpy import uic
 
 from core.module import Connector, StatusVar
-from core.util.units import get_unit_prefix_dict
 from gui.colordefs import ColorScaleInferno
-from gui.colordefs import QudiPalettePale as palette
 from gui.guibase import GUIBase
 from gui.guiutils import ColorBar
 from qtpy import QtCore
