+++ logic/fitmethods/sinemethods.py
@@ -23,7 +23,7 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 
 import numpy as np
 from lmfit.models import Model
-from core.util.units import compute_ft
+from core.util.math import compute_ft
 
 
 ################################################################################
